[CFE] Add 'double import' lint
Add a lint to discourage importing the same file more than once.
The idea being that it was probably done by mistake (e.g. automatically
by an editor).
This is done in a number of places and they are currently added to the
status file.

Change-Id: I16bc9c04c5a607f87ead1fb34922a7edb408c3da
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/113985
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>