Record request / response statistics.
If following flags are specified for DAS, in IntelliJ in Registry as
"dart.server.additional.arguments":

--new-analysis-driver-log=file:/Users/scheglov/tmp/das.log
--include-protocol-to-driver-log

then abbreviated requests, and request execution statistics is
written to the file. We also write abbreviated notifications.

The goal is to be able to see what work was requested, what was
analyzed, and identify why some requests took long to get responses.


R=brianwilkerson@google.com

Change-Id: I98e99a73f10f2842c5efb350189a9675f29f45df
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/113800
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>