[vm/mirrors] ClassMirror stop getting methods from mixin class
The problem is that a method from a mixin class was instantiated from a subclass of its mixin application.
Because _instantiator was passed from subclass to its superclass. When getter of declarations are called, it tries to get members from its mixin class. But instantiator were from subclass. This leads to a mismatch in type arguments.

The solution is to compute members from mixin appliation instead of mixin class.

Bug: https://github.com/dart-lang/sdk/issues/37360
Change-Id: Ie23f5c907209a8639cb83b45371a31f0f19cf858
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/113740
Commit-Queue: Zichang Guo <zichangguo@google.com>
Reviewed-by: Ryan Macnak <rmacnak@google.com>