Migration: fix some tests of assertion behavior.
These tests were passing for the wrong reasons: they were checking
that an edge was absent, but they were looking for the wrong edge.
Fixed the tests to look for the right edge, and added counterexamples
to demonstrate that the edge is found under the circumstances where it
is expected.

Also, the test test_if_conditional_control_flow_after() contained a
TODO comment that has since been resolved.  Deleted the TODO comment
and added a test demonstrating that the functionality now works.

Change-Id: I1442421401a54245a1299b643ccfbc1a71416d60
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/114520
Reviewed-by: Dan Rubel <danrubel@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>