Flow analysis: shift responsibility for handling final fall-through to FlowAnalysis.
Dart allows the last switch case to "fall through" the end of the
switch statement.  To account for this, we previously required the
caller of FlowAnalysis to calling `handleBreak` at the end of the last
switch case.  (In the analyzer implementation we actually only did
this if the last case was a `default` case, but in practice there was
no observable difference).

This CL shifts the responsibility to FlowAnalysis to handle this
scenario; the caller only needs to call `handleBreak` if there is an
actual break statement present.

Change-Id: Ia0d7abb21d78205f1050873e28b30cddc5fbe344
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/114547
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>