[vm, bytecode] Don't write Bytecode::var_descriptors_ in snapshots.
This matches the behavior for Code::var_descriptors and causes the field to effectively be nulled when writing a snapshot. It may be lazily recomputed by the reader. Fixes various app-jit tests in bytecode mode that causes the var descriptors to be generating by printing stack traces.

Change-Id: I04fddaf46508050eeab8af63d5c6eb1e1b7fa51f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/115000
Reviewed-by: Alexander Markov <alexmarkov@google.com>
Reviewed-by: Régis Crelier <regis@google.com>
Commit-Queue: Ryan Macnak <rmacnak@google.com>