[infra] Ignore newly detected flakes in try runs
Previously, approve_results.dart only looked at flaky.json from the
baseline build to determine whether to ignore a flake. But, if the flake
occurs in the try build and is detected as such by deflaking, flaky.json
might not contain data about the flake. Therefore, approve_results.dart
needs to also inspect the results data for the flaky bit.

Change-Id: I51f4c44a80684414c2c50d2460aed3f90eadaf12
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/116341
Reviewed-by: Jonas Termansen <sortie@google.com>