[vm/fuzzer] re-enable control flow collections
Rationale:
Let's stress test this feature again. Note
that this will increase the number of skipped
tests quite a bit since the CFE complains
about a lot of the generated constructs.
It is still TBD to find out if this is
in the grammar or in the FE.

https://github.com/dart-lang/sdk/issues/38231

Change-Id: Iec4bcb2aa9039f5b01293f82e3e156cc2e35414f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/116663
Reviewed-by: Felicitas Hetzelt <felih@google.com>
Reviewed-by: Ben Konyi <bkonyi@google.com>
Commit-Queue: Aart Bik <ajcbik@google.com>