[pkg/vm] Train kernel service differently
When the kernel service is asked to train ("--train") one gives it a script
to train on (by compiling the script).
In our build scripts that is dart2js - a relatively big application that
exercise many parts of the compilation pipeline. That is all fine and good,
but some parts are only executed once (e.g. creating the UriTranslator etc).
In an attempt to train these parts more this change additionally
compiles a simple hello world script 10 times whenever it's asked to
train on anything.

On my desktop there is no statistical significant change in runtime of
a hello world script, but on a ODROID-C2 system where I trained it
manually on dart2js before and after this change I get these runtime
changes:

        -0.0269167 +/- 0.00493467
        -3.65054% +/- 0.669259%

i.e. ~26 ms faster (from a median of 0.735 seconds to 0.708 seconds).

This CL furthermore changes the way we print via the kernel service in
verbose mode, as the current way generally prints ~5M integers to a
text buffer, then uses the first 256 characters of that and prints it to
terminal. This is here changed to only print the first ~256 characters
thus making the verbose enabled runtime more similar to the verbose
disabled runtime.

Change-Id: Id53b059025adc3517e99e56eaef438a6093b674a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/117300
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>