[vm] Fix all wasm exception memory leaks
This involved refactoring so that all the throws happen in the top level native entry functions, and then making sure that all the native objects are accounted for in those locations. I ran the wasm tests under asan and verified they used to leak but now don't.

Bug: https://github.com/dart-lang/sdk/issues/37882
Change-Id: I9991ff6b9f8af1fa2f335c73ecf11aba29ea8c72
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/117443
Commit-Queue: Liam Appelbe <liama@google.com>
Reviewed-by: Ryan Macnak <rmacnak@google.com>