[cfe] Remove NullabilityBuilder.legacy
TL;DR: Use cases of 'legacy' constructor are covered by the 'omitted'
constructor.

The initial idea behind the removed constructor was to use it on all
types in opted-out libraries, as if the programmer had a way to specify
which type is legacy.  That didn't prevent from building the 'omitted'
NullabilityBuilder into 'legacy' nullability sometimes.  In some cases
we still had to use the 'omitted' constructor because we didn't have the
access to the library object yet, and then, then the
NullabilityBuilder.build method is invoked, and the library object is
passed to it, we build 'legacy' nullability.

Change-Id: I975fa93e23ee55e93db5f3e642a97ad6901b06c1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/118568
Reviewed-by: Johnni Winther <johnniwinther@google.com>