[vm] Switch vm-kernel-precomp-mac-debug-simarm_x64 to vm-kernel-precomp-linux-debug-simarm_x64.
The reason is that we can test Assembly snapshot generation on Linux but not on Mac.

We have had several bugs propagate to Flutter end-users due to our lack of testing of
AOT Assembly snapshots on ARM (SIMARM_X64 in particular), causing reverts of SDK commits
and disrupting the Dart -> Flutter roll.

For example, see:
  - https://github.com/flutter/flutter/issues/40114
  - https://github.com/flutter/flutter/issues/41228

Change-Id: Ided8116e8230e58ffd350447b2e1a22676bd0506
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/118584
Reviewed-by: Alexander Thomas <athom@google.com>
Commit-Queue: Samir Jindel <sjindel@google.com>