Add support for details and implement some initial details for regions
Details provide a way for clients to provide navigation support. I'm
imagining that the description could be used as link text.

The deatils are currently more of a data dump than information. I need
to see more migration results in order to figure out when and how to
pare down the data into actionable information.

We might want to capture all of the navigation targets as we build the
info so that we can create anchors in the HTML.

Change-Id: Ie48622ce7d935d3c028ea42e51cfc8da590db149
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/118908
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>