[ VM / Service ] Add --object-id-ring-size to unblock UX studies for memory profiling
Rationale:
This flag is temporary and will be removed once UX studies are
completed in the next month or so. These studies are currently blocked
as DevTools has not yet migrated to use the new HeapSnapshot format and
will be unable to for at least a few weeks, and the current memory profiling
page is running into a significant number of sentinel IDs during long
sessions.

Removal is tracked here and is set for the first half of November: https://github.com/dart-lang/sdk/issues/38535

Change-Id: Ia8c1f9c65efd6690443ba9d4f41254e79091a16a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/118471
Commit-Queue: Ben Konyi <bkonyi@google.com>
Reviewed-by: Siva Annamalai <asiva@google.com>