Resolve initializers of final class fields in classes with const constructors.
ResolvedAstPrinter is used to print otherwise invisible properties of
AST nodes, such as elements and types. In this case - the static type
of the constant ListLiteral, that must get the static type, and this
type must be inherited from the context.

This solves (a surprisingly rare) breakage in google3 when I ran a
presubmit to enable stronger runtimeType constant checks using
TypeSystem.

R=brianwilkerson@google.com, paulberry@google.com

Change-Id: I9f50fb7edef344677e3b3b1ac27b136326cb2c40
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/119240
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>